파일 분리
Card Class 파일 분리
